// src/grpc/proto/plugin/grpc_stdio.ts
export enum StdioData_Channel {
  INVALID = 0,
  STDOUT = 1,
  STDERR = 2,
  UNRECOGNIZED = -1,
}